+++ test/testers/fully-connected.h
@@ -105,7 +105,7 @@ public:
 	}
 
 	void testOutput() const {
-		const uint_fast32_t seed = std::chrono::system_clock::now().time_since_epoch().count();
+		const uint_fast32_t seed = uint_fast32_t(std::chrono::system_clock::now().time_since_epoch().count());
 		auto rng = std::bind(std::uniform_real_distribution<float>(), std::mt19937(seed));
 
 		std::vector<float> input(batchSize() * inputChannels());
@@ -137,7 +137,7 @@ public:
 	void testInferenceF32() const {
 		ASSERT_EQ(1, batchSize());
 
-		const uint_fast32_t seed = std::chrono::system_clock::now().time_since_epoch().count();
+		const uint_fast32_t seed = uint_fast32_t(std::chrono::system_clock::now().time_since_epoch().count());
 		auto rng = std::bind(std::uniform_real_distribution<float>(), std::mt19937(seed));
 
 		std::vector<float> input(inputChannels());
